+++ BullishorBust/Frontend/.env.example
@@ -1,4 +1,4 @@
-EXPO_PUBLIC_BACKEND_URL=http://192.168.1.100:3000
+EXPO_PUBLIC_BACKEND_URL=http://localhost:3000/api
 ALPACA_API_KEY=PKN4ICO3WECXSLDGXCHC
 ALPACA_SECRET_KEY=PwJAEwLnLnsf7qAVvFutE8VIMgsAgvi7PMkMcCca
 ALPACA_BASE_URL=https://paper-api.alpaca.markets/v2
